% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.elo.run.R
\name{as.matrix.elo.run}
\alias{as.matrix.elo.run}
\title{as.matrix.elo.run}
\usage{
\method{as.matrix}{elo.run}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"elo.run"}.}

\item{...}{Other arguments (Not in use at this time).}
}
\description{
Convert an object from \code{\link{elo.run}} into a matrix of running Elos.
}
