% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.elo.run.R
\name{auc.elo.run}
\alias{auc.elo.run}
\title{Calculate AUC on an \code{elo.run} object}
\usage{
\method{auc}{elo.run}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{elo.run}}.}

\item{...}{Other arguments (not used at this time).}
}
\value{
The AUC of the predicted Elo probabilities and the actual win results.
}
\description{
Calculate AUC on an \code{elo.run} object
}
\references{
Adapted from code here:
  \url{https://stat.ethz.ch/pipermail/r-help/2005-September/079872.html}
}
\seealso{
\code{pROC::\link[pROC]{auc}}, \code{\link{elo.run}}.
}
