% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{elo.accuracy}
\alias{elo.accuracy}
\alias{accuracy}
\alias{accuracy.elo.run}
\alias{accuracy.elo.glm}
\alias{accuracy.elo.running}
\alias{accuracy.elo.markovchain}
\alias{accuracy.elo.winpct}
\alias{accuracy.elo.colley}
\title{Calculate accuracy}
\usage{
accuracy(object, subset, ...)

\method{accuracy}{elo.run}(object, subset, ...)

\method{accuracy}{elo.glm}(object, subset, ...)

\method{accuracy}{elo.running}(object, subset, running = TRUE, ...)

\method{accuracy}{elo.markovchain}(object, subset, ...)

\method{accuracy}{elo.winpct}(object, subset, ...)

\method{accuracy}{elo.colley}(object, subset, ...)
}
\arguments{
\item{object}{An object}

\item{subset}{(optional) A vector of indices on which to calculate accuracy.}

\item{running}{logical, denoting whether to use the running predicted values.}
}
\description{
Calculate the accuracy (percent of cases where the favored team wins)
for a given model.
}
