% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.beta.R
\name{elo.beta}
\alias{elo.beta}
\title{Compute win probabilities based on the beta distribution}
\usage{
elo.beta(
  formula,
  data,
  na.action,
  subset,
  initial.alphas = NULL,
  initial.betas = initial.alphas,
  ...
)
}
\arguments{
\item{formula}{A formula. See \link[=formula.specials]{the help page for formulas} for details.}

\item{data}{A \code{data.frame} in which to look for objects in \code{formula}.}

\item{na.action}{A function which indicates what should happen when the data contain NAs.}

\item{subset}{An optional vector specifying a subset of observations.}

\item{initial.alphas, initial.betas}{Initial values. Default is 1 (akin to a uniform distribution).}

\item{...}{Other arguments (not in use at this time).}
}
\description{
Compute win probabilities based on the beta distribution
}
