% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.model.frame.R
\name{elo.model.frame}
\alias{elo.model.frame}
\title{Interpret formulas in \code{elo} functions}
\usage{
elo.model.frame(formula, data, na.action, subset, ..., envir)
}
\arguments{
\item{formula}{A formula. See "details", below.}

\item{data}{A \code{data.frame} in which to look for objects in \code{formula}.}

\item{na.action}{A function which indicates what should happen when the data contain NAs.}

\item{subset}{An optional vector specifying a subset of observations.}

\item{...}{Other arguments (not in use at this time).}

\item{envir}{An environment in which to evaluate \code{\link[stats]{model.frame}}. Almost always
\code{\link{parent.frame}}.}
}
\description{
A helper function to create the model frame for many \code{elo} functions.
}
\seealso{
\code{\link{elo.run}}, \code{\link{elo.calc}}, \code{\link{elo.prob}}
}
