% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.R
\name{elo.mov}
\alias{elo.mov}
\alias{mov}
\title{Create a "margin of victory" column}
\usage{
mov(score.A, score.B = 0)
}
\arguments{
\item{score.A}{Numeric; the score of the first team. Alternatively, this can
be a pre-computed margin of victory which will get compared to 0.}

\item{score.B}{Numeric; the score of the second team; default is 0, in case
\code{score.A} is already a margin of victory..}
}
\value{
An object with class \code{"elo.mov"}, denoting \code{score.A} = \code{score.B}.
}
\description{
Create a "margin of victory" based on two teams' scores
}
\examples{
mov(12, 10)
mov(10, 10)
mov(10, 12)
}
\seealso{
\code{\link{score}}
}
