% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{elo.mse}
\alias{elo.mse}
\alias{mse}
\alias{mse.elo.glm}
\alias{mse.elo.run}
\title{Calculate the mean square error}
\usage{
mse(object, subset, ...)

\method{mse}{elo.glm}(object, subset, ...)

\method{mse}{elo.run}(object, subset, ...)
}
\arguments{
\item{object}{An object}

\item{subset}{(optional) A vector of indices on which to calculate the MSE.}

\item{...}{Other arguments (not in use at this time).}
}
\description{
Calculate the mean square error
}
\details{
Even though logistic regressions don't use the MSE on the y=0/1 scale, it can still be informative.
}
