% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\docType{package}
\name{elo}
\alias{elo}
\alias{elo-package}
\title{The Elo Package}
\description{
An implementation of Elo ratings for general use in 'R'.
}
\section{Functions}{


Below are listed some of the functions available in \code{elo}:

\code{\link{elo.calc}}: Calculate Elos for a series of matches.
}

\section{Data}{


\code{\link{tournament}}: Mock data for examples.
}

\examples{
library(elo)
}
\references{
Elo, A. E. 1978. The Rating of Chess Players, Past and Present. New York: Arco.
}
