% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favored.R
\name{favored.elo}
\alias{favored.elo}
\alias{favored}
\alias{favored.elo.run}
\alias{favored.elo.glm}
\alias{favored.elo.running}
\alias{favored.elo.markovchain}
\alias{favored.elo.winpct}
\alias{favored.elo.colley}
\alias{favored.default}
\title{Classify teams that are favored to win}
\usage{
favored(x, ..., subset = TRUE)

\method{favored}{elo.run}(x, ..., subset = TRUE)

\method{favored}{elo.glm}(x, ..., subset = TRUE)

\method{favored}{elo.running}(x, running = TRUE, discard.skipped = FALSE, ..., subset = TRUE)

\method{favored}{elo.markovchain}(x, ..., subset = TRUE)

\method{favored}{elo.winpct}(x, ..., subset = TRUE)

\method{favored}{elo.colley}(x, ..., subset = TRUE)

\method{favored}{default}(x, p.A, ...)
}
\arguments{
\item{x}{An object from \code{\link{elo.run}} or \code{\link{elo.glm}}, or for the default method
a vector representing wins.A.}

\item{...}{Other arguments (not used at this time).}

\item{subset}{(optional) A vector of indices on which to calculate}

\item{running}{logical, denoting whether to use the running predicted values.}

\item{discard.skipped}{Logical, denoting whether to ignore the skipped observations in the calculation}

\item{p.A}{A vector of predicted win probabilities.}
}
\description{
Classify teams that are favored to win
}
