% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.train.R, R/fit.elo.run.R
\name{fit.elo.run}
\alias{fit.elo.run}
\title{Train models with a gridsearch}
\usage{
fit.elo.run(formula, data, optimfun = "mse", gridsearch, ...)

fit.elo.run(formula, data, optimfun = "mse", gridsearch, ...)
}
\arguments{
\item{formula}{A formula. See \link[=formula.specials]{the help page for formulas} for details.}

\item{data}{A \code{data.frame} in which to look for objects in \code{formula}.}

\item{optimfun}{A string refering to an optimization function to train the
model on. See details.}

\item{gridsearch}{A dataframe of values to be used for optimization.
See details.}

\item{...}{Other arguments (not used at this time).}
}
\value{
An elo object optimized by the given optimization function

An elo object optimized by the given optimization function
}
\description{
Takes in the arguments of the given elo function along with an optimization
function and a grid of parameters to find the best combination of the
parameters.

Takes in the arguments of the given elo.run function along with an optimization
function and a grid of parameters to find the best combination of the
parameters.
}
\details{
\code{optimfun} refers to one of the following optimization functions
built into the elo package. 'mse' or 'brier' will train on minimizing the
mean squared error of the model, 'accuracy' will train on maximizing the
accuracy of the mode, 'calibration' will train on getting the calibration score
of the model as close to 1 as possible, and 'auc' will train on maximizing the
auc of the model.

\code{gridsearch} is a dataframe of parameter values that \code{fit.elo} will
search through to optimize. The dataframe's column names correspond to the
parameters to be optimized within the formula (i.e. if the term
\code{adjust(team.home,hfa)} is included, to optimize hfa you will need a
column named 'hfa' with plausible values in the grid search dataframe.
A column named "k" must be included. If you wish to optimize other parameters
while holding k constant, simply set k to the same value for all rows of the
grid search.
}
\examples{
data(tournament)
params = expand.grid(k = seq(1,250,by=5),
                     hfa = seq(0,100,by=5))
fit.elo.run(score(points.Home, points.Visitor) ~ adjust(team.Home, hfa) + team.Visitor,
            data = tournament,
            optimfun = 'mse',
            gridsearch = params)

params = list(hfa = seq(0,100,by=5))
fit.elo.glm(score(points.Home, points.Visitor) ~ adjust(team.Home, hfa) + team.Visitor,
            data = tournament,
            optimfun = 'mse',
            gridsearch = params)
data(tournament)
params = expand.grid(k = seq(1,250,by=5),
                     hfa = seq(0,100,by=5))
fit.elo.run(score(points.Home, points.Visitor) ~ adjust(team.Home, hfa) + team.Visitor,
            data = tournament,
            optimfun = 'mse',
            gridsearch = params)
}
