% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.elo}
\alias{fitted.elo}
\alias{fitted.elo.run}
\alias{residuals.elo.run}
\alias{fitted.elo.running}
\alias{fitted.elo.glm}
\alias{fitted.elo.markovchain}
\alias{fitted.elo.winpct}
\alias{fitted.elo.colley}
\title{Extract model values}
\usage{
\method{fitted}{elo.run}(object, ...)

\method{residuals}{elo.run}(object, ...)

\method{fitted}{elo.running}(object, running = TRUE, ...)

\method{fitted}{elo.glm}(object, ...)

\method{fitted}{elo.markovchain}(object, ...)

\method{fitted}{elo.winpct}(object, ...)

\method{fitted}{elo.colley}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Other arguments}

\item{running}{logical, denoting whether to use the running predicted values.}
}
\description{
Extract model values from \code{elo} functions.
}
