% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.elo.colley.R
\name{predict.elo.colley}
\alias{predict.elo.colley}
\title{Make Predictions on an \code{elo.colley} Object}
\usage{
\method{predict}{elo.colley}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{"\link{elo.colley}"}.}

\item{newdata}{A new dataset containing the same variables as the call
that made \code{object}. If missing, the predicted win probabilities from
\code{object} will be returned.}

\item{...}{Other arguments.}
}
\value{
A vector of win probabilities.
}
\description{
Make Predictions on an \code{elo.colley} Object
}
\examples{
data(tournament)
t1 <- head(tournament, -3)
t2 <- tail(tournament, 3)
results <- elo.colley(score(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = t1,
  subset = points.Home != points.Visitor)
predict(results)
predict(results, newdata = t2)
}
\seealso{
\code{\link{predict.elo.running}}
}
