% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.teams.R
\name{rank.teams}
\alias{rank.teams}
\alias{rank.teams.elo.run}
\alias{rank.teams.elo.run.regressed}
\alias{rank.teams.elo.glm}
\alias{rank.teams.elo.markovchain}
\alias{rank.teams.elo.winpct}
\alias{rank.teams.elo.colley}
\title{Rank teams}
\usage{
rank.teams(object, ties.method = "min", ...)

\method{rank.teams}{elo.run}(object, ties.method = "min", ...)

\method{rank.teams}{elo.run.regressed}(object, ties.method = "min", regressed = FALSE, ...)

\method{rank.teams}{elo.glm}(object, ties.method = "min", ...)

\method{rank.teams}{elo.markovchain}(object, ties.method = "min", ...)

\method{rank.teams}{elo.winpct}(object, ties.method = "min", ...)

\method{rank.teams}{elo.colley}(object, ties.method = "min", ...)
}
\arguments{
\item{object}{An object.}

\item{ties.method}{Passed to \code{\link{rank}}.}

\item{...}{Other arguments}

\item{regressed}{Passed to \code{\link{final.elos}}.}
}
\description{
Extract the rankings from Elo objects.
}
