% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Create a 1/0/0.5 win "indicator"}
\usage{
score(score.A, score.B)
}
\arguments{
\item{score.A}{Numeric; the score of the first team (whose wins are to be denoted by 1).}

\item{score.B}{Numeric; the score of the second team (whose wins are to be denoted by 0).}
}
\value{
A vector containing 0, 1, and 0.5 (for ties).
}
\description{
Create a 1/0/0.5 win "indicator" based on two teams' scores.
}
\examples{
score(12, 10)
score(10, 10)
score(10, 12)
}
