% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.elo.colley.R
\name{summary.elo.colley}
\alias{summary.elo.colley}
\title{Summarize an \code{elo.colley} Object}
\usage{
\method{summary}{elo.colley}(object, ...)
}
\arguments{
\item{object}{An object resulting from \code{\link{elo.colley}}.}

\item{...}{Other arguments}
}
\value{
A summary of \code{object}.
}
\description{
Summarize an \code{elo.colley} Object
}
\examples{
co <- elo.colley(score(points.Home, points.Visitor) ~ team.Home + team.Visitor,
  data = tournament, subset = points.Home != points.Visitor)
summary(co)
}
\seealso{
\code{\link{elo.colley}}, \code{\link{favored}}, \code{\link{auc.elo.colley}}, \code{\link{mse}}
}
