% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.elo.markovchain.R
\name{summary.elo.markovchain}
\alias{summary.elo.markovchain}
\title{Summarize an \code{elo.markovchain} Object}
\usage{
\method{summary}{elo.markovchain}(object, ...)
}
\arguments{
\item{object}{An object resulting from \code{\link{elo.markovchain}}.}

\item{...}{Other arguments}
}
\value{
A summary of \code{object}.
}
\description{
Summarize an \code{elo.markovchain} Object
}
\examples{
mc <- elo.markovchain(score(points.Home, points.Visitor) ~ team.Home + team.Visitor,
  data = tournament, subset = points.Home != points.Visitor, k = 0.7)
summary(mc)
}
\seealso{
\code{\link{elo.markovchain}}, \code{\link{favored}}, \code{\link{auc.elo.markovchain}}, \code{\link{mse}}
}
