% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.elo.run.R
\name{summary.elo.run}
\alias{summary.elo.run}
\title{Summarize an \code{elo.run} Object}
\usage{
\method{summary}{elo.run}(object, ...)
}
\arguments{
\item{object}{An object resulting from \code{\link{elo.run}}.}

\item{...}{Other arguments}
}
\value{
A summary of \code{object}.
}
\description{
Summarize an \code{elo.run} Object
}
\examples{
summary(elo.run(score(points.Home, points.Visitor) ~ team.Home + team.Visitor,
  data = tournament, k = 20))
}
\seealso{
\code{\link{elo.run}}, \code{\link{favored}}, \code{\link{auc.elo.run}}, \code{\link{mse}}
}
