% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.elo.winpct.R
\name{summary.elo.winpct}
\alias{summary.elo.winpct}
\title{Summarize an \code{elo.winpct} Object}
\usage{
\method{summary}{elo.winpct}(object, ...)
}
\arguments{
\item{object}{An object resulting from \code{\link{elo.winpct}}.}

\item{...}{Other arguments}
}
\value{
A summary of \code{object}.
}
\description{
Summarize an \code{elo.winpct} Object
}
\examples{
wp <- elo.winpct(score(points.Home, points.Visitor) ~ team.Home + team.Visitor,
  data = tournament, subset = points.Home != points.Visitor, k = 0.7)
summary(wp)
}
\seealso{
\code{\link{elo.winpct}}, \code{\link{favored}}, \code{\link{auc.elo.winpct}}, \code{\link{mse}}
}
