% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.multiteam.R
\name{tournament.multiteam}
\alias{tournament.multiteam}
\title{\code{tournament.multiteam}: Mock data for examples}
\format{
A data frame with 56 observations on the following 4 variables:
  \describe{
    \item{\code{week}}{Week Number}
    \item{\code{half}}{The half of the season in which the match was played}
    \item{\code{Place_1}}{The first-place team}
    \item{\code{Place_2}}{The second-place team}
    \item{\code{Place_3}}{The third-place team}
    \item{\code{Place_4}}{The fourth-place team}
  }
}
\description{
A fake dataset containing results from "animal-ball" matches.
}
