% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{tournament}
\alias{tournament}
\title{\code{tournament}: Mock data for examples}
\format{A data frame with 56 observations on the following 4 variables:
  \describe{
    \item{\code{team.Home}}{The home team for the match}
    \item{\code{team.Visitor}}{The visiting team for the match}
    \item{\code{points.Home}}{Number of points scored by the home team}
    \item{\code{points.Visitor}}{Number of points scored by the visiting team}
    \item{\code{week}}{Week Number}
    \item{\code{half}}{The half of the season in which the match was played}
  }}
\description{
A fake dataset containing results from "animal-ball" matches.
}
\examples{
data(tournament)
str(tournament)
}
