% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.calc.R
\name{elo.calc}
\alias{elo.calc}
\alias{elo.calc}
\alias{elo.calc.default}
\alias{elo.calc.formula}
\title{Elo functions}
\usage{
elo.calc(elo.A, ...)

\method{elo.calc}{default}(elo.A, elo.B, wins.A, k, ...)

\method{elo.calc}{formula}(formula, data, na.action, subset, k = NULL, ...)
}
\arguments{
\item{elo.A, elo.B}{Numeric vectors of elo scores.}

\item{...}{Other arguments (not in use at this time).}

\item{wins.A}{Numeric vector of wins by team A.}

\item{k}{Numeric vector or scalar of k-values.}

\item{formula}{A formula. See "details", below.}

\item{data}{A \code{data.frame} in which to look for objects in \code{formula}.}

\item{na.action}{A function which indicates what should happen when the data contain NAs.}

\item{subset}{An optional vector specifying a subset of observations.}
}
\description{
\code{elo.prob} calculates the probability that team A beats team B.
\code{elo.update} calculates the update value for a given Elo matchup, and is used in
\code{elo.calc}, which reports the post-update Elo values.
}
\details{
These are all vectorized.


  Originally, I was going to script these in Rcpp, but the performance benefits are only realized
  when the vectors get to be ~1000 elements long.
}
\examples{
elo.calc(c(1500, 1500), c(1500, 1600), c(1, 0), k = 20)
}
\seealso{
\code{\link{elo.prob}}, \code{\link{elo.update}}
}
