% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.model.frame.R
\name{elo.model.frame}
\alias{elo.model.frame}
\title{Interpret formulas in \code{elo} functions}
\usage{
elo.model.frame(formula, data, na.action, subset, k = NULL, ...,
  required.vars = "teams")
}
\arguments{
\item{formula}{A formula. See "details", below.}

\item{data}{A \code{data.frame} in which to look for objects in \code{formula}.}

\item{na.action}{A function which indicates what should happen when the data contain NAs.}

\item{subset}{An optional vector specifying a subset of observations.}

\item{k}{A constant k-value (or a vector, where appropriate).}

\item{...}{Other arguments (not in use at this time).}

\item{required.vars}{One or more of \code{c("wins", "teams", "k")}, denoting which variables
are required to appear in the final model.frame..}
}
\description{
A helper function to create the \code{model.frame} for many \code{elo} functions.
}
\details{
With the exception of the formula in \code{\link{elo.run}},
  \code{formula} is usually of the form \code{wins.A ~ elo.A + elo.B}, where \code{elo.A} and \code{elo.B}
  are vectors of Elos, and \code{wins.A} is between 0 and 1,
  denoting whether team A (Elo A) won or lost (or something between).

\code{formula} accepts two special functions in it. \code{k()} allows for complicated Elo updates. For
  constant Elo updates, use the \code{k = } argument instead of this special function.
  \code{adjust()} allows for Elos to be adjusted for, e.g., home-field advantage. The second argument
  to this function can be a scalar or vector of appropriate length.
}
\seealso{
\code{\link{elo.run}}, \code{\link{elo.calc}}, \code{\link{elo.prob}}
}
