% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.prob.R
\name{elo.prob}
\alias{elo.prob}
\alias{elo.prob}
\alias{elo.prob.default}
\alias{elo.prob.formula}
\title{Elo functions}
\usage{
elo.prob(elo.A, ...)

\method{elo.prob}{default}(elo.A, elo.B, ...)

\method{elo.prob}{formula}(formula, data, na.action, subset, ...)
}
\arguments{
\item{elo.A}{Numeric vectors of elo scores.}

\item{...}{Other arguments (not in use at this time).}

\item{elo.B}{Numeric vectors of elo scores.}

\item{formula}{A formula. See the "details" section of \code{\link{elo.model.frame}}.}

\item{data}{A \code{data.frame} in which to look for objects in \code{formula}.}

\item{na.action}{A function which indicates what should happen when the data contain NAs.}

\item{subset}{An optional vector specifying a subset of observations.}
}
\description{
Calculate the probability that team A beats team B. This is vectorized.
}
\examples{
elo.prob(1500, 1500)
elo.prob(c(1500, 1500), c(1500, 1600))
}
\seealso{
\code{\link{elo.update}}, \code{\link{elo.calc}}
}
