% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.run.R
\name{elo.run}
\alias{elo.run}
\title{Calculate Elos for a series of matches}
\usage{
elo.run(formula, data, k, na.action, subset, initial.elo = NULL, ...)
}
\arguments{
\item{formula}{A formula of the form \code{wins.A ~ team.A + team.B}, where \code{team.A} and \code{team.B}
are character vectors or factors denoting which two teams played, and \code{wins.A} is between 0 and 1,
denoting whether team A won or lost (or something between). The teams can be adjusted for other variables.
See "details", below.}

\item{data}{A \code{data.frame} in which to look for objects in \code{formula} and \code{k}.}

\item{k}{The k-value, specified as a vector. This allows for complicated Elo updates. See "details", below.}

\item{na.action}{A function which indicates what should happen when the data contain NAs.}

\item{subset}{An optional vector specifying a subset of observations.}

\item{initial.elo}{An optional named vector containing initial Elo ratings for all teams in \code{formula}.}

\item{...}{Other arguments (not used at this time).}
}
\description{
Calculate Elos for a series of matches.
}
\examples{
data(tournament)
tournament$k <- 20
elo.calc(score(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = tournament, k = k)
}
