% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.elo.glm.R
\name{summary.elo.glm}
\alias{summary.elo.glm}
\title{Summarize an \code{elo.glm} Object}
\usage{
\method{summary}{elo.glm}(object, ...)
}
\arguments{
\item{object}{An object resulting from \code{\link{elo.glm}}.}

\item{...}{Other arguments}
}
\value{
A summary of \code{object}.
}
\description{
Summarize an \code{elo.glm} Object
}
\examples{
summary(elo.glm(score(points.Home, points.Visitor) ~ team.Home + team.Visitor,
  data = tournament))
}
\seealso{
\code{\link{elo.glm}}, \code{\link{favored}}, \code{\link{auc.elo.run}}, \code{\link{mse}}
}
